TmapAPI 호출(미완)

+++ src/main/java/iriro/saferoute/dto/RoutePointDto.java
@@ -6,12 +6,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
 public class RoutePointDto {
-    private double lat; // 위도
-    private double lng; // 경도
-    private int sequence; // 경로 순서 -> 몇 번째 지점이 위험구역인지 판단
+    private BigDecimal lat; // 위도
+    private BigDecimal lng; // 경도
+    private Integer sequence; // 경로 순서 -> 몇 번째 지점이 위험구역인지 판단
 }
